moved getToken to helpers
